G-formula Analysis

Mediation analysis for time varying mediator, estimation based on g-formula. Output contains total effect, #' natural direct effect and natural indirect effect for mediation or regular g-formula. data.frame will be returned.

Gformula(
  data,
  id.var,
  base.vars,
  exposure,
  time.var,
  models,
  intervention = NULL,
  ref_int = 0,
  init.recode = NULL,
  in.recode = NULL,
  out.recode = NULL,
  mc.sample = 10000,
  R = 500,
  ncores = 1L
)

Arguments

data
Data set to be sued

id.var
ID variable per subject.

base.vars
A vector of time fixed baseline variables.

exposure
Intervention/Exposure variable

time.var
Time variable.

models
A list of models for the G-formula, including exposure model, covariate model (if any), mediator model (if any), outcome model or censoring model (if any). See details in spec_model. The order appeared in the list should reflect the temporal ordering of the variables, in another way data generation process. The model will be evaluated in this process.

intervention
A named list with a value of intervention on exposure. if kept as NULL (default), the natural intervention course will be calculated. eg: list(natural = NULL, always = c(1, 1, 1), never = c(0, 0, 0))

ref_int
Integer or character string denoting the intervention to be used as the reference for calculating the risk ratio and risk difference. 0 or "natural" denotes the natural course, while subsequent integers denote user-specified interventions in the order that they are named in intervention. Or if the character is the name of the intervention. The default is 0.

init.recode
optional, recoding of variables done at the beginning of the Monte Carlo loop. Needed for operations initialize baseline variables. This is executed at beginning of the Monte Carlo g-formula, executed only once at time 0.

in.recode
optional, On the fly recoding of variables done before the Monte Carlo loop starts. Needed to do any kind of functional forms for entry times. This is executed at each start of the Monte Carlo g-formula time steps

out.recode
optional, On the fly recoding of variables done at the end of the Monte Carlo loop. Needed for operations like counting the number of days with a treatment or creating lagged variables. This is executed at each end of the Monte Carlo g-formula time steps.

mc.sample
Sample size of Monte Carlo simulation.

R
The number of bootstrap replicates, default is 500.

ncores
integer: number of processes to be used in parallel operation: typically one would chose this to the number of available CPUs. Parallel computation for bootstrap will be applied if ncores larger than 1.

is.survival
Is the data survival data, default is FALSE.

Note

Not that final outcome must be the same for in all rows per subject. If the dataset is survival settings, records after the interested outcome must be deleted. The function it self do some data manipulation internally. Please prepare the data as longitudinal format.

References

Lin, S. H., Young, J. G., Logan, R., & VanderWeele, T. J. (2017). Mediation analysis for a survival outcome with time-varying exposures, mediators, and confounders. Statistics in medicine, 36(26), 4153-4166. DOI:10.1002/sim.7426 Zheng, W., & van der Laan, M. (2017). Longitudinal mediation analysis with time-varying mediators and exposures, with application to survival outcomes. Journal of causal inference, 5(2). DOI:10.1515/jci-2016-0006