Main calculation function

Main calculation function

.gformula(
  data,
  id.var,
  base.vars,
  time.var,
  exposure,
  models,
  intervention,
  in.recode,
  out.recode,
  init.recode,
  mediation_type = c(NA, "N", "I"),
  mc.sample = 10000,
  return_fitted = FALSE
)

Arguments

data
Data set to be sued

id.var
ID variable per subject.

base.vars
A vector of time fixed baseline variables.

time.var
Time variable.

exposure
Intervention/Exposure variable

models
A list of models for the G-formula, including exposure model, covariate model (if any), mediator model (if any), outcome model or censoring model (if any). See details in spec_model. The order appeared in the list should reflect the temporal ordering of the variables, in another way data generation process. The model will be evaluated in this process.

intervention
A named list with a value of intervention on exposure. if kept as NULL (default), the natural intervention course will be calculated. eg: list(natural = NULL, always = c(1, 1, 1), never = c(0, 0, 0))

in.recode
optional, On the fly recoding of variables done before the Monte Carlo loop starts. Needed to do any kind of functional forms for entry times. This is executed at each start of the Monte Carlo g-formula time steps

out.recode
optional, On the fly recoding of variables done at the end of the Monte Carlo loop. Needed for operations like counting the number of days with a treatment or creating lagged variables. This is executed at each end of the Monte Carlo g-formula time steps.

init.recode
optional, recoding of variables done at the beginning of the Monte Carlo loop. Needed for operations initialize baseline variables. This is executed at beginning of the Monte Carlo g-formula, executed only once at time 0.

mediation_type
Type of the mediation effect.

mc.sample
Sample size of Monte Carlo simulation.

return_fitted
Return the fitted model (default is FALSE).