Bone marrow transplant data

The study population arose from a multicenter trial of leukemia patients and comprises 137 individuals prepared for bone marrow transplants under a radiation-free regimen at four medical centers.

gvhd

Format

A data frame with 108714 rows and 29 variables:

id

unique id of subject

age

baseline age

male

male indicator

cmv

cytomegalovirus immune status (yes or no)

all

leukemia type (acute lymphocytic = 1 or acute myeloid leukemia = 0)

wait

wait time from leukemia diagnosis to transplantation

day

days since bone marrow transplant

d

indicator of death (1= yes, 0=no)

gvhd

indicator of GvHD (1= yes, 0=no)

daysgvhd

number of days since onset of GvHD

daysnogvhd

number of days without GvHD

gvhdm1

lagged GvHD

relapse

relapse indicator

relapsem1

lagged relapse indicator

daysrelapse

number of days since relapse

daysnorelapse

number of days relapse-free

platnorm

normal platelet levels (1=patient has relapsed or reached normal platelet count, 0=not in relapse or below normal platelets)

platnormm1

lagged normal platelet levels

daysplatnorm

time spent reaching normal platelet levels

daysnoplatnorm

time spent without reaching normal platelet levels

censlost

indicator of censoring due to loss-to-follow up (1=yes, 0=no)

References

Keil et al. (2014) Epidemiology Epidemiology, 25(6), 889–897 (PubMed)