Model specification for G-formula

Specify regression model for the time-varying variables. This is will be used for g-formula. An unevaluated model will be created and can further pass to Gformula.

spec_model(
  formula,
  subset = NULL,
  recode = NULL,
  var_type = c("normal", "binary", "categorical", "custom"),
  mod_type = c("exposure", "covariate", "mediator", "outcome", "censor", "survival"),
  custom_fit = NULL,
  custom_sim = NULL,
  ...
)

Arguments

formula
an object of class formula: a symbolic description of the model to be fitted. Formula for specified models passed to the model. Must be contained within the input `data.frame` when initialized. This will passed to the model fitting function.

subset
an optional vector specifying a subset of observations to be used in the fitting process and passed to the model fitting function.

recode
An Optional string vector. Recoding variable before including variables in the model. Strings will be evaluated before the fitting the model.

var_type
A character strings specifying the "type" of response variable of the model. The possible values are: "binary", "normal", "categorical", and "custom".

mod_type
Model type. Exposure model ("exposure"), covariate model ("covariate"), mediator model ("mediator"), outcome model ("outcome"), mediator model ("survival") or censoring model ("censor").

custom_fit
Custom model fitting function. The var_type should be set to "custom" to pass the `custom_fit`.

custom_sim
Custom simulation function for the model, see sim_value. Only two parameters will be pass to this function during the evaluation, the fitted model object and the data to be used in the prediction. The first argument of the function must be the fitted model object and the second must be the data to be used in the simulation. If the custom simulation function was not provided, the variable will be simulated using the normal distribution.

...
Other parameters passed to the model fitting function or custom_fit.

Value

The list returned is:

call

An unevaluated model.

subset

A string vector passed by the parameter subset.

recode

A string vector passed by the parameter recode.

var_type

A character string passed by the parameter var_type.

mod_type

A character string passed by the parameter mod_type.

custom_sim

A symbolic customized simulation function name, passed by the parameter custom_sim.

Details

This function will be used to create an unevaluated model for the g-formula.

See also